move: 각 topic event package 이동
