+++ src/main/java/com/project/yogerOrder/cart/event/consumer/CartEventConsumer.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Component;
 import com.project.yogerOrder.cart.dto.request.DeleteProductFromCartRequestDTO;
 import com.project.yogerOrder.cart.service.CartService;
 import com.project.yogerOrder.global.config.KafkaConfig;
-import com.project.yogerOrder.order.config.OrderTopic;
+import com.project.yogerOrder.order.event.config.OrderTopic;
 import com.project.yogerOrder.order.dto.request.OrderItemRequestDTO;
 import com.project.yogerOrder.order.event.OrderCompletedEvent;
